feat: hook up a new backend
feat: hook up a new backend

The provided API would sometimes return search results with `shelf` set to the wrong value for books we have on a shelf. This made keeping books in sync across views weird; syncing up the shelves meant we had to have all of our shelved books downloaded in order to set the right value for search results. This is no longer the case.